requirements.txt
Flask~=3.1.2
openai~=1.109.1
pydantic~=2.12.5
guardrails-ai~=0.7.0 #guardrails
#arize~=7.51.1
#pandas~=2.3.3
bandit~=1.9.2
boto3~=1.42.3
requests~=2.32.5
numpy~=2.3.5
python-dotenv~=1.2.1
#arize-otel~=0.11.0
#opentelemetry-api~=1.39.0
#opentelemetry-sdk~= 1.39.0
arize-phoenix-otel~=0.14.0
openinference-instrumentation-openai~=0.1.41
matplotlib~=3.10.7

